BAckend/Temp/requirements.txt
Flask~=3.1.0
pycryptodome~=3.22.0
